requirements.txt
torch
transformers
Pillow
matplotlib
inflect
gTTS
gradio
requests
